Move community list into separate component
