+++ src/pages/posts/CommunitiesPage.tsx
@@ -1,124 +1,13 @@
-import {
-  IonHeader,
-  IonIcon,
-  IonItem,
-  IonItemDivider,
-  IonItemGroup,
-  IonLabel,
-  IonList,
-  IonPage,
-  IonTitle,
-  IonToolbar,
-} from "@ionic/react";
-import AppContent from "../../features/shared/AppContent";
-import { useParams } from "react-router";
-import { useAppDispatch, useAppSelector } from "../../store";
-import { getHandle } from "../../helpers/lemmy";
-import { home, library, people, star } from "ionicons/icons";
-import styled from "@emotion/styled";
-import { pullAllBy, sortBy, uniqBy } from "lodash";
-import { notEmpty } from "../../helpers/array";
-import { Fragment, useMemo, useRef } from "react";
+import { useRef } from "react";
+import CommunitiesList from "../../features/community/list/CommunitiesList";
 import { useSetActivePage } from "../../features/auth/AppContext";
-import { useBuildGeneralBrowseLink } from "../../helpers/routes";
-import ItemIcon from "../../features/labels/img/ItemIcon";
-import { jwtSelector } from "../../features/auth/authSlice";
-import { Community } from "lemmy-js-client";
-import { ActionButton } from "../../features/post/actions/ActionButton";
-import {
-  addFavorite,
-  removeFavorite,
-} from "../../features/community/communitySlice";
-import { css } from "@emotion/react";
-
-const SubIcon = styled(IonIcon)<{ color: string }>`
-  border-radius: 50%;
-  padding: 6px;
-  width: 1rem;
-  height: 1rem;
-
-  background: ${({ color }) => color};
-  --ion-color-base: white;
-`;
-
-const Content = styled.div`
-  margin: 0.7rem 0;
-
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-
-  aside {
-    margin-top: 0.2rem;
-    color: var(--ion-color-medium);
-    font-size: 0.8em;
-  }
-`;
+import { IonHeader, IonPage, IonTitle, IonToolbar } from "@ionic/react";
+import AppContent from "../../features/shared/AppContent";
 
 export default function CommunitiesPage() {
-  const buildGeneralBrowseLink = useBuildGeneralBrowseLink();
-  const { actor } = useParams<{ actor: string }>();
-  const jwt = useAppSelector(jwtSelector);
   const pageRef = useRef();
 
-  const follows = useAppSelector((state) => state.auth.site?.my_user?.follows);
-
-  const communityByHandle = useAppSelector(
-    (state) => state.community.communityByHandle
-  );
-
-  const favorites = useAppSelector((state) => state.community.favorites);
-
   useSetActivePage(pageRef.current);
-
-  const communities = useMemo(() => {
-    const communities = uniqBy(
-      [
-        ...(follows || []).map((f) => f.community),
-        ...Object.values(communityByHandle).map((c) => c?.community),
-      ].filter(notEmpty),
-      "id"
-    );
-
-    pullAllBy(
-      communities,
-      Object.values(communityByHandle)
-        .filter((response) => response?.subscribed === "NotSubscribed")
-        .map((c) => c?.community),
-      "id"
-    );
-
-    return communities;
-  }, [follows, communityByHandle]);
-
-  const favoritesAsCommunitiesIfFound = useMemo(
-    () =>
-      favorites.map(
-        (community) =>
-          communities.find((c) => community === getHandle(c)) || community
-      ),
-    [communities, favorites]
-  );
-
-  const communitiesGroupedByLetter = useMemo(() => {
-    const alphabeticallySortedCommunities = sortBy(communities, (c) =>
-      c.name.toLowerCase()
-    );
-
-    return Object.entries(
-      alphabeticallySortedCommunities.reduce<Record<string, Community[]>>(
-        (acc, community) => {
-          const firstLetter = community.name[0].toUpperCase();
-          if (!acc[firstLetter]) {
-            acc[firstLetter] = [];
-          }
-          acc[firstLetter].push(community);
-          return acc;
-        },
-        {}
-      )
-    );
-  }, [communities]);
 
   return (
     <IonPage ref={pageRef}>
@@ -128,136 +17,8 @@ export default function CommunitiesPage() {
         </IonToolbar>
       </IonHeader>
       <AppContent scrollY>
-        <IonList>
-          <IonItemGroup>
-            {jwt && (
-              <IonItem routerLink={buildGeneralBrowseLink(`/home`)}>
-                <Content>
-                  <SubIcon icon={home} color="red" />
-                  <div>
-                    Home
-                    <aside>Posts from subscriptions</aside>
-                  </div>
-                </Content>
-              </IonItem>
-            )}
-            <IonItem routerLink={buildGeneralBrowseLink(`/all`)}>
-              <Content>
-                <SubIcon icon={library} color="#009dff" />
-                <div>
-                  All<aside>Posts across all federated communities</aside>
-                </div>
-              </Content>
-            </IonItem>
-            <IonItem routerLink={buildGeneralBrowseLink(`/local`)} lines="none">
-              <Content>
-                <SubIcon icon={people} color="#00f100" />
-                <div>
-                  Local<aside>Posts from communities on {actor}</aside>
-                </div>
-              </Content>
-            </IonItem>
-          </IonItemGroup>
-
-          {favoritesAsCommunitiesIfFound.length > 0 && (
-            <>
-              <IonItemGroup>
-                <IonItemDivider>
-                  <IonLabel>Favorites</IonLabel>
-                </IonItemDivider>
-              </IonItemGroup>
-
-              {favoritesAsCommunitiesIfFound.map((favorite) =>
-                typeof favorite === "string" ? (
-                  <IonItem
-                    key={favorite}
-                    routerLink={buildGeneralBrowseLink(`/c/${favorite}`)}
-                  >
-                    <Content>
-                      <ItemIcon item={favorite} size={28} />
-                      {favorite}
-                    </Content>
-                  </IonItem>
-                ) : (
-                  <RenderCommunity
-                    key={favorite.id}
-                    community={favorite}
-                    favorites={favorites}
-                  />
-                )
-              )}
-            </>
-          )}
-
-          {communitiesGroupedByLetter.map(([letter, communities]) => (
-            <Fragment key={letter}>
-              <IonItemGroup>
-                <IonItemDivider>
-                  <IonLabel>{letter}</IonLabel>
-                </IonItemDivider>
-              </IonItemGroup>
-              {communities.map((community) => (
-                <RenderCommunity
-                  key={community.id}
-                  community={community}
-                  favorites={favorites}
-                />
-              ))}
-            </Fragment>
-          ))}
-        </IonList>
+        <CommunitiesList />
       </AppContent>
     </IonPage>
-  );
-}
-
-function RenderCommunity({
-  community,
-  favorites,
-}: {
-  community: Community;
-  favorites?: string[];
-}) {
-  const buildGeneralBrowseLink = useBuildGeneralBrowseLink();
-  const dispatch = useAppDispatch();
-
-  const isFavorite = useMemo(
-    () => favorites?.includes(getHandle(community)) ?? false,
-    [favorites, community]
-  );
-
-  return (
-    <IonItem
-      key={community.id}
-      routerLink={buildGeneralBrowseLink(`/c/${getHandle(community)}`)}
-    >
-      <Content>
-        <ItemIcon item={community} size={28} />
-        {getHandle(community)}
-      </Content>
-      <ActionButton
-        slot="end"
-        onClick={(e) => {
-          e.stopPropagation();
-          e.preventDefault();
-
-          const handle = getHandle(community);
-
-          if (!isFavorite) {
-            dispatch(addFavorite(handle));
-          } else {
-            dispatch(removeFavorite(handle));
-          }
-        }}
-      >
-        <IonIcon
-          icon={star}
-          color={isFavorite ? "primary" : "light"}
-          css={css`
-            font-size: 24px;
-          `}
-        />
-      </ActionButton>
-    </IonItem>
   );
 }
